requirements.txt
numpy==2.1.3
pandas==2.2.3
matplotlib==3.9.2
opencv-python==4.10.0.84
Pillow==11.0.0
tqdm==4.66.6
pyyaml==6.0.2
